indentation.rkt : stop spin on unbreakable lines
Long scribble lines which have no natural break-point like:

@author+email["mrxxxyyyyyyyesqIII"]{xxx.yyyyyyy@foocorp.co.uk}

would cause reindent-paragraph to spin. In fact it would insert newlines between " and m, which is just inappropriate.

Change made to insert-break-function to ensure that quoted strings are not broken.
These are literal strings which IMHO shouldn't be tampered with; unlike the {} enclosed at-strings.