indentation.rkt : stop spin on unbreakable lines
A bug in insert-break-func caused a spin on lines which can only be broken to the right of width. See the subsubsubsubsection example.

posi iterates from (+ start width) in the first half of the function, but only from width at the bottom. changed to add start to the return value.